refactor: general elements file per component
